requirement.txt
sentence-transformers==2.2.2
langchain
flask
sentence-transformers
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
-e .